+++ frontend/src/common/components/TagList.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { EmbyProps } from '~/emby/components/embyProps';
+import './TagList.css';
 
 export type TagListProps = {
     Items: string[]
@@ -10,7 +11,7 @@ export const TagList: React.FC<TagListProps> = props => {
     return (
         <div className='tag-list'>
             {props.Items.map(x =>
-                <div className='emby-button'>{x}</div>
+                <div className='raised emby-button tag-item'>{x}</div>
             )}
         </div>
     )